# api-contracts/openapi/v1/attribution.yaml
openapi: 3.1.0
info:
  title: Skeldir Attribution API
  version: 2.0.0
  description: |
    Attribution Intelligence API - Realtime Revenue Counter and Attribution Data.
    Provides verified revenue metrics with statistical confidence scoring.
  contact:
    name: Skeldir Engineering
    email: engineering@skeldir.com
    url: https://skeldir.com

servers:
  - url: http://localhost:4011
    description: Prism Mock Server (Attribution)

tags:
  - name: Attribution
    description: Attribution and revenue tracking endpoints
  - name: LLM Explanation
    description: LLM-powered natural language explanations for attribution data

security:
  - bearerAuth: []

paths:
  /api/attribution/revenue/realtime:
    get:
      summary: Get Realtime Revenue Counter
      description: |
        Polls realtime revenue metrics with 30-second cache TTL.
        Supports ETag-based caching with If-None-Match header for bandwidth optimization.
      deprecated: true
      operationId: getRealtimeRevenue
      tags:
        - Attribution
      security:
        - bearerAuth: []
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
        - $ref: './_common/base.yaml#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Revenue counter data returned successfully
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
              description: Request correlation ID echoed back
            ETag:
              schema:
                type: string
              description: Entity tag for cache validation
            Cache-Control:
              schema:
                type: string
              description: Cache control directives
              example: max-age=30
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeRevenueCounter'
              example:
                total_revenue: 125430.50
                event_count: 1247
                last_updated: '2025-11-26T14:32:00Z'
                data_freshness_seconds: 5
                verified: true
                tenant_id: 00000000-0000-0000-0000-000000000000
                confidence_score: 0.94
                upgrade_notice: null
        '304':
          description: Not Modified - ETag matches, use cached data
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
            ETag:
              schema:
                type: string
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '429':
          $ref: './_common/base.yaml#/components/responses/RateLimitError'
        '500':
          $ref: './_common/base.yaml#/components/responses/ServerError'

  /api/attribution/channels:
    get:
      summary: Get Channel Attribution Data
      description: Returns attribution data broken down by marketing channel
      operationId: getChannelAttribution
      tags:
        - Attribution
      security:
        - bearerAuth: []
      parameters:
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date for attribution window
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date for attribution window
      responses:
        '200':
          description: Channel attribution data
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                type: object
                required:
                  - channels
                  - total_revenue
                  - date_range
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChannelAttribution'
                  total_revenue:
                    type: number
                    format: double
                    minimum: 0
                  date_range:
                    type: object
                    properties:
                      start:
                        type: string
                        format: date
                      end:
                        type: string
                        format: date
              example:
                channels:
                  - channel_name: Meta
                    revenue: 45250.75
                    conversion_count: 185
                    confidence_score: 0.92
                    spend: 18000.00
                    roas: 2.51
                  - channel_name: Google
                    revenue: 37840.10
                    conversion_count: 142
                    confidence_score: 0.89
                    spend: 15000.00
                    roas: 2.52
                total_revenue: 83090.85
                date_range:
                  start: '2025-11-01'
                  end: '2025-11-15'
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'

  /api/attribution/explain/{entity_type}/{entity_id}:
    get:
      summary: Get natural language explanation for attribution entities
      description: |
        Fast RAG endpoint for interactive explanations. Centaur architecture:
        retrieves deterministic attribution data, synthesizes via LLM.
        Target latency: <500ms p95.
      operationId: explainAttributionEntity
      tags:
        - Attribution
        - LLM Explanation
      parameters:
        - name: entity_type
          in: path
          required: true
          schema:
            type: string
            enum:
              - attribution_score
              - channel_performance
              - reconciliation_discrepancy
          description: Type of entity to explain
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the entity
        - $ref: './_common/base.yaml#/components/parameters/CorrelationId'
        - $ref: './_common/base.yaml#/components/parameters/Authorization'
      responses:
        '200':
          description: Explanation with citations and confidence scoring
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
              description: Request correlation ID echoed back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributionExplanation'
              example:
                explanation: "Facebook ROAS of $3.20 is outperforming Google ROAS of $2.10 because last-touch conversions skew toward paid social during the Black Friday campaign. Confidence is high given 90 days of stable conversion data."
                citations:
                  - metric: facebook_roas
                    value: 3.20
                    source: attribution_results.channel_roas
                  - metric: google_roas
                    value: 2.10
                    source: attribution_results.channel_roas
                confidence: high
                generated_at: '2025-11-26T14:30:00Z'
                cost_cents: 1.8
        '401':
          $ref: './_common/base.yaml#/components/responses/UnauthorizedError'
        '404':
          $ref: './_common/base.yaml#/components/responses/NotFoundError'
        '500':
          $ref: './_common/base.yaml#/components/responses/ServerError'
      x-latency-requirement: 500ms
      x-cost-budget: 0.02
      x-complexity-score: 5

components:
  schemas:
    RealtimeRevenueCounter:
      type: object
      required:
        - total_revenue
        - event_count
        - last_updated
        - data_freshness_seconds
        - verified
        - tenant_id
      properties:
        total_revenue:
          type: number
          format: double
          minimum: 0
          description: Total verified revenue amount in USD
          example: 125430.50
        event_count:
          type: integer
          description: Number of revenue events tracked
          minimum: 0
          example: 1247
        last_updated:
          type: string
          format: date-time
          description: ISO 8601 timestamp of last data update
          example: '2025-11-26T14:32:00Z'
        data_freshness_seconds:
          type: integer
          description: Seconds since data was last refreshed
          minimum: 0
          example: 5
        verified:
          type: boolean
          description: Whether revenue is verified through payment reconciliation
          example: true
        tenant_id:
          type: string
          format: uuid
          description: Tenant identifier associated with the response
          example: '00000000-0000-0000-0000-000000000000'
        confidence_score:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Statistical confidence score (0-1) for attribution accuracy
          example: 0.94
        upgrade_notice:
          type: string
          nullable: true
          description: Optional message about upgrading for more features
          example: Upgrade to Pro for historical analytics

    ChannelAttribution:
      type: object
      required:
        - channel_name
        - revenue
        - conversion_count
        - confidence_score
      properties:
        channel_name:
          type: string
          enum: [Meta, Google, TikTok, LinkedIn, Organic, Direct, Email, Referral]
          description: Marketing channel name
        revenue:
          type: number
          format: double
          minimum: 0
          description: Attributed revenue for this channel
        conversion_count:
          type: integer
          description: Number of conversions attributed to this channel
        confidence_score:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Statistical confidence in attribution accuracy
        spend:
          type: number
          format: float
          description: Marketing spend on this channel
        roas:
          type: number
          format: float
          description: Return on Ad Spend (revenue/spend)

    AttributionExplanation:
      type: object
      required:
        - explanation
        - citations
        - confidence
        - generated_at
      properties:
        explanation:
          type: string
          minLength: 150
          maxLength: 500
          description: Natural language explanation of attribution result
          example: "Your Facebook ROAS of $3.20 indicates strong performance because you have 90 days of stable conversion data with narrow confidence bounds ($2.95-$3.45)."
        citations:
          type: array
          items:
            type: object
            required:
              - metric
              - value
              - source
            properties:
              metric:
                type: string
                example: facebook_roas
              value:
                type: number
                example: 3.20
              source:
                type: string
                description: Database table.column reference
                example: attribution_results.channel_roas
        confidence:
          type: string
          enum: [high, medium, low]
          description: Explanation confidence based on data quality
        generated_at:
          type: string
          format: date-time
        cost_cents:
          type: number
          format: float
          description: LLM API cost for this explanation
          example: 1.8
          x-internal: true

  securitySchemes:
    bearerAuth:
      $ref: './_common/base.yaml#/components/securitySchemes/bearerAuth'